requirements.txt
langchain
langchain-community
langchain-huggingface
chromadb
sentence-transformers
pypdf
langchain-google-genai
python-dotenv
google-ai-generativelanguage==0.6.6
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.7.2
googleapis-common-protos==1.72.0
pymupdf